% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_decisions}
\alias{picosat_decisions}
\alias{picosat_decisions.picosat_solution}
\title{Number of decisions}
\usage{
picosat_decisions(x)

\method{picosat_decisions}{picosat_solution}(x)
}
\arguments{
\item{x}{a solution from the solver}
}
\value{
length 1 int vector
}
\description{
Number of decisions
}
