% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_propagations}
\alias{picosat_propagations}
\alias{picosat_propagations.picosat_solution}
\title{Number of propagations}
\usage{
picosat_propagations(x)

\method{picosat_propagations}{picosat_solution}(x)
}
\arguments{
\item{x}{a solution from the solver}
}
\value{
length 1 int vector
}
\description{
Number of propagations
}
