% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_seconds}
\alias{picosat_seconds}
\alias{picosat_seconds.picosat_solution}
\title{Solution time in seconds}
\usage{
picosat_seconds(x)

\method{picosat_seconds}{picosat_solution}(x)
}
\arguments{
\item{x}{a solution from the solver}
}
\value{
length 1 numeric vector in seconds
}
\description{
Solution time in seconds
}
