% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_solution_status}
\alias{picosat_solution_status}
\alias{picosat_solution_status.picosat_solution}
\title{Get the solution status}
\usage{
picosat_solution_status(x)

\method{picosat_solution_status}{picosat_solution}(x)
}
\arguments{
\item{x}{a solution from the solver}
}
\value{
character either PICOSAT_SATISFIABLE,
  PICOSAT_UNSATISFIABLE or PICOSAT_UNKNOWN
}
\description{
Get the solution status
}
