% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_variables}
\alias{picosat_variables}
\alias{picosat_variables.picosat_solution}
\title{The number of variables in a model}
\usage{
picosat_variables(x)

\method{picosat_variables}{picosat_solution}(x)
}
\arguments{
\item{x}{an object}
}
\value{
an integer vector of length 1
}
\description{
The number of variables in a model
}
