% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_added_original_clauses}
\alias{picosat_added_original_clauses}
\title{The number of original clauses}
\usage{
picosat_added_original_clauses(x)
}
\arguments{
\item{x}{a picosat solution object}
}
\value{
an integer vector of length 1
}
\description{
The number of original clauses
}
