% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_decisions}
\alias{picosat_decisions}
\title{The number of decisions during a search}
\usage{
picosat_decisions(x)
}
\arguments{
\item{x}{a picosat solution object}
}
\value{
an integer vector of length 1
}
\description{
The number of decisions during a search
}
