% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_propagations}
\alias{picosat_propagations}
\alias{picosat_propagations.picosat_solution}
\title{The number of propagations during a search}
\usage{
picosat_propagations(x)

\method{picosat_propagations}{picosat_solution}(x)
}
\arguments{
\item{x}{an object}
}
\value{
an integer vector of length 1
}
\description{
The number of propagations during a search
}
