% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_propagations}
\alias{picosat_propagations}
\title{The number of propagations during a search}
\usage{
picosat_propagations(x)
}
\arguments{
\item{x}{a picosat solution object}
}
\value{
an integer vector of length 1
}
\description{
The number of propagations during a search
}
