% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_sat}
\alias{picosat_sat}
\title{Solve SAT problems with the PicoSAT solver}
\usage{
picosat_sat(formula, assumptions = integer(0), verbosity_level = 0L)
}
\arguments{
\item{formula}{a list of integer vectors. Each vector is a clause.
Each integer identifies a literal. No element must be 0.
Negative integers are negated literals.}

\item{assumptions}{an optional integer vector. Assumptions are preset values for literals in your formula.
Each element correspond to a literal.
Negative literals are FALSE, positive TRUE.}

\item{verbosity_level}{either 0, 1, 2 where 2 is the most verbose log level.}
}
\value{
a data.frame with two columns, variable and value. In case the solution status
  is not PICOSAT_SATISFIABLE the resulting data.frame has 0 rows.
}
\description{
The solver takes a formula in conjunctive normal form and finds
a satisfiable assignment of the
literals or returns that the formula is not satisfiable.
}
\examples{
formula <- list(
 c(-1, 2), # 1 => 2
 c(-2, 3)  # 2 => 3
)
picosat_sat(formula, 1) # we set 1 to TRUE
}
