% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_sat}
\alias{picosat_sat}
\title{Solve SAT problems with the PicoSAT solver}
\usage{
picosat_sat(cnf, assumptions = integer(0), verbosity_level = 0L)
}
\arguments{
\item{verbosity_level}{either 0, 1, 2 where 2 is the most verbose log level.}

\item{clause}{a list of integer vectors. Each vector is a clause.
Each integer identifies a literal. No element must be 0.
Negative integers are negated literals.}

\item{assumption}{an optional integer vector. Assumptions are preset values for literals in your formula.
Each element correspond to a literal.
Negative literals are FALSE, positive TRUE.}
}
\description{
The solver takes a formula in conjunctive normal form and finds
a satisfiable assignment of the
literals or returns that the formula is not satisfiable.
}
\examples{
formula <- list(
 c(-1, 2), # 1 => 2
 c(-2, 3)  # 2 => 3
)
picosat_sat(formula, 1) # we set 1 to TRUE
}
