% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_seconds}
\alias{picosat_seconds}
\title{Time spent in `picosat_sat`}
\usage{
picosat_seconds(x)
}
\arguments{
\item{x}{a picosat solution object}
}
\value{
a numeric vector of length 1
}
\description{
Time spent in `picosat_sat`
}
