% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_variables}
\alias{picosat_variables}
\title{The number of variables in a model}
\usage{
picosat_variables(x)
}
\arguments{
\item{x}{a picosat solution object}
}
\value{
an integer vector of length 1
}
\description{
The number of variables in a model
}
