% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_visits}
\alias{picosat_visits}
\alias{picosat_visits.picosat_solution}
\title{Number of visits}
\usage{
picosat_visits(x)

\method{picosat_visits}{picosat_solution}(x)
}
\arguments{
\item{x}{a solution from the solver}
}
\value{
length 1 int vector
}
\description{
Number of visits
}
