% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpicosat.R
\name{picosat_visits}
\alias{picosat_visits}
\title{The number of visits during a search}
\usage{
picosat_visits(x)
}
\arguments{
\item{x}{a picosat solution object}
}
\value{
an integer vector of length 1
}
\description{
The number of visits during a search
}
